+++ allennlp/tests/commands/find_learning_rate_test.py
@@ -10,9 +10,10 @@ from allennlp.data import Vocabulary, DataIterator
 from allennlp.models import Model
 from allennlp.common.checks import ConfigurationError
 from allennlp.common.testing import AllenNlpTestCase
-from allennlp.commands.train import datasets_from_params, Trainer
+from allennlp.commands.train import Trainer
 from allennlp.commands.find_learning_rate import search_learning_rate, \
     find_learning_rate_from_args, find_learning_rate_model, FindLearningRate
+from allennlp.training.util import datasets_from_params
 
 
 class TestFindLearningRate(AllenNlpTestCase):
