+++ allennlp/training/util.py
@@ -1,0 +1,240 @@
+"""
+Helper functions for Trainers
+"""
+from typing import Any, Union, Dict, Iterable
+import datetime
+import logging
+import os
+import shutil
+
+import torch
+from tensorboardX import SummaryWriter
+
+from allennlp.common.checks import ConfigurationError
+from allennlp.common.params import Params
+from allennlp.data.dataset_readers import DatasetReader
+from allennlp.data.instance import Instance
+from allennlp.models.archival import archive_model, CONFIG_NAME
+
+logger = logging.getLogger(__name__)
+
+
+def sparse_clip_norm(parameters, max_norm, norm_type=2) -> float:
+    """Clips gradient norm of an iterable of parameters.
+
+    The norm is computed over all gradients together, as if they were
+    concatenated into a single vector. Gradients are modified in-place.
+    Supports sparse gradients.
+
+    Parameters
+    ----------
+    parameters : ``(Iterable[torch.Tensor])``
+        An iterable of Tensors that will have gradients normalized.
+    max_norm : ``float``
+        The max norm of the gradients.
+    norm_type : ``float``
+        The type of the used p-norm. Can be ``'inf'`` for infinity norm.
+
+    Returns
+    -------
+    Total norm of the parameters (viewed as a single vector).
+    """
+    # pylint: disable=invalid-name,protected-access
+    parameters = list(filter(lambda p: p.grad is not None, parameters))
+    max_norm = float(max_norm)
+    norm_type = float(norm_type)
+    if norm_type == float('inf'):
+        total_norm = max(p.grad.data.abs().max() for p in parameters)
+    else:
+        total_norm = 0
+        for p in parameters:
+            if p.grad.is_sparse:
+                # need to coalesce the repeated indices before finding norm
+                grad = p.grad.data.coalesce()
+                param_norm = grad._values().norm(norm_type)
+            else:
+                param_norm = p.grad.data.norm(norm_type)
+            total_norm += param_norm ** norm_type
+        total_norm = total_norm ** (1. / norm_type)
+    clip_coef = max_norm / (total_norm + 1e-6)
+    if clip_coef < 1:
+        for p in parameters:
+            if p.grad.is_sparse:
+                p.grad.data._values().mul_(clip_coef)
+            else:
+                p.grad.data.mul_(clip_coef)
+    return total_norm
+
+
+def move_optimizer_to_cuda(optimizer):
+    """
+    Move the optimizer state to GPU, if necessary.
+    After calling, any parameter specific state in the optimizer
+    will be located on the same device as the parameter.
+    """
+    for param_group in optimizer.param_groups:
+        for param in param_group['params']:
+            if param.is_cuda:
+                param_state = optimizer.state[param]
+                for k in param_state.keys():
+                    if isinstance(param_state[k], torch.Tensor):
+                        param_state[k] = param_state[k].cuda(device=param.get_device())
+
+
+def get_batch_size(batch: Union[Dict, torch.Tensor]) -> int:
+    """
+    Returns the size of the batch dimension. Assumes a well-formed batch,
+    returns 0 otherwise.
+    """
+    if isinstance(batch, torch.Tensor):
+        return batch.size(0) # type: ignore
+    elif isinstance(batch, Dict):
+        return get_batch_size(next(iter(batch.values())))
+    else:
+        return 0
+
+
+class TensorboardWriter:
+    """
+    Wraps a pair of ``SummaryWriter`` instances but is a no-op if they're ``None``.
+    Allows Tensorboard logging without always checking for Nones first.
+    """
+    def __init__(self, train_log: SummaryWriter = None, validation_log: SummaryWriter = None) -> None:
+        self._train_log = train_log
+        self._validation_log = validation_log
+
+    @staticmethod
+    def _item(value: Any):
+        if hasattr(value, 'item'):
+            val = value.item()
+        else:
+            val = value
+        return val
+
+    def add_train_scalar(self, name: str, value: float, global_step: int) -> None:
+        # get the scalar
+        if self._train_log is not None:
+            self._train_log.add_scalar(name, self._item(value), global_step)
+
+    def add_train_histogram(self, name: str, values: torch.Tensor, global_step: int) -> None:
+        if self._train_log is not None:
+            if isinstance(values, torch.Tensor):
+                values_to_write = values.cpu().data.numpy().flatten()
+                self._train_log.add_histogram(name, values_to_write, global_step)
+
+    def add_validation_scalar(self, name: str, value: float, global_step: int) -> None:
+
+        if self._validation_log is not None:
+            self._validation_log.add_scalar(name, self._item(value), global_step)
+
+
+def time_to_str(timestamp: int) -> str:
+    """
+    Convert seconds past Epoch to human readable string.
+    """
+    datetimestamp = datetime.datetime.fromtimestamp(timestamp)
+    return '{:04d}-{:02d}-{:02d}-{:02d}-{:02d}-{:02d}'.format(
+            datetimestamp.year, datetimestamp.month, datetimestamp.day,
+            datetimestamp.hour, datetimestamp.minute, datetimestamp.second
+    )
+
+
+def str_to_time(time_str: str) -> datetime.datetime:
+    """
+    Convert human readable string to datetime.datetime.
+    """
+    pieces: Any = [int(piece) for piece in time_str.split('-')]
+    return datetime.datetime(*pieces)
+
+
+def datasets_from_params(params: Params) -> Dict[str, Iterable[Instance]]:
+    """
+    Loads the train / validation / test datasets specified by the config.
+    """
+    train_reader = DatasetReader.from_params(params.pop('dataset_reader'))
+    if "validation_dataset_reader" in params:
+        logger.info("Using a separate dataset reader to load validation and test data.")
+        validation_reader = DatasetReader.from_params(params.pop('validation_dataset_reader'))
+    else:
+        validation_reader = train_reader
+
+    datasets: Dict[str, Iterable[Instance]] = {}
+    for reader, name, key in zip([train_reader, validation_reader],
+                                 ['train', 'validation', 'test'],
+                                 ['train_data_path', 'validation_data_path', 'test_data_path']):
+        data_path = params.pop(key, None)
+        if data_path is not None:
+            logger.info(f"reading {name} data from {key} {data_path}")
+
+            datasets[name] = reader.read(data_path)
+
+    return datasets
+
+def create_serialization_dir(
+        params: Params,
+        serialization_dir: str,
+        recover: bool,
+        force: bool) -> None:
+    """
+    This function creates the serialization directory if it doesn't exist.  If it already exists
+    and is non-empty, then it verifies that we're recovering from a training with an identical configuration.
+
+    Parameters
+    ----------
+    params: ``Params``
+        A parameter object specifying an AllenNLP Experiment.
+    serialization_dir: ``str``
+        The directory in which to save results and logs.
+    recover: ``bool``
+        If ``True``, we will try to recover from an existing serialization directory, and crash if
+        the directory doesn't exist, or doesn't match the configuration we're given.
+    force: ``bool``
+        If ``True``, we will overwrite the serialization directory if it already exists.
+    """
+    if recover and force:
+        raise ConfigurationError("Illegal arguments: both force and recover are true.")
+
+    if os.path.exists(serialization_dir) and force:
+        shutil.rmtree(serialization_dir)
+
+    if os.path.exists(serialization_dir) and os.listdir(serialization_dir):
+        if not recover:
+            raise ConfigurationError(f"Serialization directory ({serialization_dir}) already exists and is "
+                                     f"not empty. Specify --recover to recover training from existing output.")
+
+        logger.info(f"Recovering from prior training at {serialization_dir}.")
+
+        recovered_config_file = os.path.join(serialization_dir, CONFIG_NAME)
+        if not os.path.exists(recovered_config_file):
+            raise ConfigurationError("The serialization directory already exists but doesn't "
+                                     "contain a config.json. You probably gave the wrong directory.")
+        else:
+            loaded_params = Params.from_file(recovered_config_file)
+
+            # Check whether any of the training configuration differs from the configuration we are
+            # resuming.  If so, warn the user that training may fail.
+            fail = False
+            flat_params = params.as_flat_dict()
+            flat_loaded = loaded_params.as_flat_dict()
+            for key in flat_params.keys() - flat_loaded.keys():
+                logger.error(f"Key '{key}' found in training configuration but not in the serialization "
+                             f"directory we're recovering from.")
+                fail = True
+            for key in flat_loaded.keys() - flat_params.keys():
+                logger.error(f"Key '{key}' found in the serialization directory we're recovering from "
+                             f"but not in the training config.")
+                fail = True
+            for key in flat_params.keys():
+                if flat_params.get(key, None) != flat_loaded.get(key, None):
+                    logger.error(f"Value for '{key}' in training configuration does not match that the value in "
+                                 f"the serialization directory we're recovering from: "
+                                 f"{flat_params[key]} != {flat_loaded[key]}")
+                    fail = True
+            if fail:
+                raise ConfigurationError("Training configuration does not match the configuration we're "
+                                         "recovering from.")
+    else:
+        if recover:
+            raise ConfigurationError(f"--recover specified but serialization_dir ({serialization_dir}) "
+                                     "does not exist.  There is nothing to recover from.")
+        os.makedirs(serialization_dir, exist_ok=True)
